% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgazetteer
\docType{data}
\name{gazetteer.2010}
\alias{gazetteer.2010}
\title{2010 Gazetteer data}
\format{A list of data.frames with appropriate names}
\source{
\url{http://www.census.gov/geo/maps-data/data/gazetteer2010.html}
}
\usage{
gazetteer.2010
}
\description{
A list of datasets containing 2010 Gazetteer data from U.S. Census Bureau.
}
\keyword{datasets}

