% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgazetteer
\docType{data}
\name{gazetteer.2013}
\alias{gazetteer.2013}
\title{2013 Gazetteer data}
\format{A list of data.frames with appropriate names}
\source{
\url{http://www.census.gov/geo/maps-data/data/gazetteer2013.html}
}
\usage{
gazetteer.2013
}
\description{
A list of datasets containing 2013 Gazetteer data from U.S. Census Bureau.
}
\keyword{datasets}

