% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgazetteer
\docType{data}
\name{gazetteer.2015}
\alias{gazetteer.2015}
\title{2015 Gazetteer data}
\format{A list of data.frames with appropriate names}
\source{
\url{http://www.census.gov/geo/maps-data/data/gazetteer2015.html}
}
\usage{
gazetteer.2015
}
\description{
A list of datasets containing 2015 Gazetteer data from U.S. Census Bureau.
}
\keyword{datasets}

