% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgazetteer.R
\docType{data}
\name{gazetteer.2016}
\alias{gazetteer.2016}
\title{2016 Gazetteer data}
\format{A list of data.frames with appropriate names}
\source{
\url{http://www.census.gov/geo/maps-data/data/gazetteer2016.html}
}
\usage{
gazetteer.2016
}
\description{
A list of datasets containing 2016 Gazetteer data from U.S. Census Bureau.
}
\keyword{datasets}
