% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgazetteer.R
\docType{data}
\name{state.areas.2010}
\alias{state.areas.2010}
\title{2010 State Gazetteer data}
\format{A data.frame with all U.S. States and Territories along with their FIPS/GNIS IDs and land and water areas.}
\source{
\url{http://www.census.gov/geo/reference/state-area.html}
}
\usage{
state.areas.2010
}
\description{
data.frame containing U.S. States and Territories
}
\keyword{datasets}

