% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgazetteer-package.r
\docType{package}
\name{usgazetteer}
\alias{usgazetteer}
\alias{usgazetteer-package}
\title{usgazetteer.}
\description{
This package allows you to access gazetteer data released by the U.S. census bureau.
}
